float to int

+++ Source/GP4Testing/PlayerSystems/WeaponManagementSystem.cpp
@@ -198,10 +198,10 @@ TMap<WeaponType, UGunComponent*> AWeaponManagementSystem::GetAcquiredWeapons()
 	return AcquiredWeapons;
 }
 
-float AWeaponManagementSystem::GetMaxAmmo()
+int AWeaponManagementSystem::GetMaxAmmo()
 {
 	UGunComponent* Weapon = nullptr;
-	float MaxAmmo = 0;
+	int MaxAmmo = 0;
 	if (AcquiredWeapons.Num() > 0)
 	{
 		Weapon = AcquiredWeapons.FindRef(EquippedWeapon);
@@ -210,10 +210,10 @@ float AWeaponManagementSystem::GetMaxAmmo()
 	return MaxAmmo;
 }
 
-float AWeaponManagementSystem::GetAmmo()
+int AWeaponManagementSystem::GetAmmo()
 {
 	UGunComponent* Weapon = nullptr;
-	float Ammo = 0;
+	int Ammo = 0;
 	if (AcquiredWeapons.Num() > 0)
 	{
 		Weapon = AcquiredWeapons.FindRef(EquippedWeapon);
@@ -222,7 +222,7 @@ float AWeaponManagementSystem::GetAmmo()
 	return Ammo;
 }
 
-bool AWeaponManagementSystem::AddAmmo(PickupType type, float newAmmo)
+bool AWeaponManagementSystem::AddAmmo(PickupType type, int newAmmo)
 {
 	UGunComponent* Weapon = nullptr;
 
@@ -245,8 +245,8 @@ bool AWeaponManagementSystem::AddAmmo(PickupType type, float newAmmo)
 		{
 			if ((Weapon->Ammo + newAmmo) > Weapon->MaxAmmo)
 			{
-				float NoAdd = (Weapon->Ammo + newAmmo) - Weapon->MaxAmmo;
-				float Add = newAmmo - NoAdd;
+				int NoAdd = (Weapon->Ammo + newAmmo) - Weapon->MaxAmmo;
+				int Add = newAmmo - NoAdd;
 				Weapon->Ammo += Add;
 				return true;
 			}
@@ -261,10 +261,10 @@ bool AWeaponManagementSystem::AddAmmo(PickupType type, float newAmmo)
 	return false;
 }
 
-float AWeaponManagementSystem::GetMaxMagazine()
+int AWeaponManagementSystem::GetMaxMagazine()
 {
 	UGunComponent* Weapon = nullptr;
-	float MaxMag = 0;
+	int MaxMag = 0;
 	if (AcquiredWeapons.Num() > 0)
 	{
 		Weapon = AcquiredWeapons.FindRef(EquippedWeapon);
@@ -273,10 +273,10 @@ float AWeaponManagementSystem::GetMaxMagazine()
 	return MaxMag;
 }
 
-float AWeaponManagementSystem::GetLoadedMagazine()
+int AWeaponManagementSystem::GetLoadedMagazine()
 {
 	UGunComponent* Weapon = nullptr;
-	float Mag = 0;
+	int Mag = 0;
 	if (AcquiredWeapons.Num() > 0)
 	{
 		Weapon = AcquiredWeapons.FindRef(EquippedWeapon);
